Chnaged from poltergist to selenium for js tests
React was not working with polterguist so i swiched to selenium
for js tests. To get it to work i had to instal xvfb on the
server. I then used the headless gem within my tests to start a
xvfb server that the selenium driver could use to run the tests.